# [ISSUE #97] Performance optimization