@RongtongJin 2021-09-26
> 我在本地 执行mvn -B package jacoco:report coveralls:report 是下面这个错误
> Failed tests:
> BatchPushTest.testBatchPushTruncate:251 expected:<4> but was:<9>
> 
> 但是 这个test 本地执行没有问题
> 
> LeaderElectorTest.testThreeServer:91 expected:<2> but was:<1>

这个应该是单测sleep时间不足导致的，我来处理一下