@Cczzzz 2021-09-26
> > check failed ,how do i fix it
> 
> It seems to have nothing to do with your code

It looks like this