@RongtongJin 2022-03-29
Hi @Cczzzz 关于第一个修改项，发现将waitForRunning(1); 改为 Thread.yield();后，即使没有消息，该线程也会将一个CPU跑满，因此考虑回退到将waitForRunning(1)